// src/types/airtable.generated.ts
export interface AirtableNotes {
  id: string;
  createdTime: string;
  fields: {
    [key: string]: any;
    /** formula field (fldQ9C8FopcgY5olR) */
    ID: string;
    /** multipleRecordLinks field (fldnNg4T43UMokhNB) */
    Lead?: string[];
    /** multipleRecordLinks field (fldV9X9o3K9xNgbka) */
    User?: string[];
    /** multilineText field (fldcN1Bb8fjT46OyO) */
    Content?: string;
    /** singleSelect field (fldEe2eW9PboYAcRS) */
    Type?: 'Riflessione' | 'Promemoria' | 'Follow-up' | 'Info Cliente';
    /** checkbox field (flduQTxMKNqGqe5i8) */
    Pinned?: boolean;
    /** createdTime field (fldY6MDn5uxeCtkBA) */
    CreatedAt?: any;
  };
}